FEATURE: Use backend configuration manager for backend settings
This used to be the frontend configuration manager, which had several
conditions the backend configuration manager didn't know. Now the
backend configuration manager brings most of them as well.
Unfortunately there's a TYPO3 core bug allowing the backend configuration
to support both, page.id and as well as tree.rootLineIds but both
stick with 0 as page so the calculated TypoScript is valid but wrong.

See: https://forge.typo3.org/issues/92120
Related: UVEXGROUP-2268